Ajustements

+++ logic/utils/Math.java
@@ -17,7 +17,7 @@ public class Math {
 	    return new Color(r, g, b);
 	}
 	
-	public static long map(long x, long in_min, long in_max, long out_min, long out_max)
+	public static float map(float x, float in_min, float in_max, float out_min, float out_max)
 	{
 	  return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 	}
